// frontend/src/types.ts
export interface ContestAnnouncement {
  id: number
  contest_id: number
  author_id: number
  author_name: string
  title: string
  content: string
  pinned: boolean
  created_at: string
  updated_at: string
}

export interface ContestQuestion {
  id: number
  contest_id: number
  asker_id: number
  asker_name: string
  content: string
  answer: string
  answerer_id?: number
  answerer_name?: string
  public: boolean
  asked_at: string
  answered_at?: string
}

export interface ContestCommunications {
  announcements: ContestAnnouncement[]
  questions: ContestQuestion[]
}